+++ Changelog.md
@@ -16,6 +16,7 @@ Bugfixes:
  * Fix internal error when popping a dynamic storage array of mappings.
  * Yul Optimizer: Fix reordering bug in connection with shifted one and mul/div-instructions in for loop conditions.
  * Scanner: Fix multi-line natspec comment parsing with triple slashes when file is encoded with CRLF instead of LF.
+ * Name Resolver: Fix wrong source location when warning on shadowed aliases in import declarations.
 
 
 ### 0.5.11 (2019-08-12)